Make UnionFileCollection cheaper
Store it either as an immutable set or
as a union view instead of using the more
memory hungry linked hash set.

We should deprecate #add, it is only called
in very few places and there is no good way
that a user can actually get a hold of a collection
that supports #add. Then this code would be even
simpler.