Remove FileResolver.resolveLater(), since resolve() is cheap now
This is a follow-up to making FileResolver faster. It is now
so cheap that resolveLater() no longer serves any purpose. The
result can no longer be different over time either, as resolve()
no longer looks at the file system.

Removing that code saves a lot of memory when repeatedly looking
up the build directory of a project (which happens a lot).